# Add optional support for memory pooling