… example/service.twirp.go
… internal/twirptest/importmapping/x/x.twirp.go
… internal/twirptest/multiple/multiple1.twirp.go
… internal/twirptest/multiple/multiple2.twirp.go
… internal/twirptest/no_package_name/no_package_name.twirp.go
… internal/twirptest/no_package_name_importer/no_package_name_importer.twirp.go
… internal/twirptest/proto/proto.twirp.go
… internal/twirptest/service.twirp.go
… internal/twirptest/service_method_same_name/service_method_same_name.twirp.go
… internal/twirptest/snake_case_names/snake_case_names.twirp.go
… internal/twirptest/source_relative/source_relative.twirp.go
… protoc-gen-twirp/generator.go