Hi all!! So we've been working on a Twirp service here at GitHub that handles pretty large ProtoBuf payloads and we found that (de)serialization can get a bit slow -- at least slow enough to show up in the profiles, which is never a good thing!

I wanted to improve this situation, but I know that you folks really care about backwards compatibility, so this is a bit of a restrained PR (no user-facing changes, no breaking changes at all and disabled by default) that enables memory pooling for the temporary buffers used to (de)serialize requests and responses.

I think the code is quite non-intrusive and the benchmark results are very significant considering the constraints, so overall I'm quite happy with the result. I would love to hear if you have any ideas to make this even less intrusive and easier to land on master.

Commit message inline for your convenience:

```
Because of the way the ProtoBuf APIs have been designed, wherein
serialization and de-serialization of messages must be performed on a
full byte slice (i.e. incremental (de)serialization is not possible),
the current version of Twirp's codegen allocates a lot of temporary
memory for each request:

- A full `bytes.Buffer` is allocated when reading the message by
`ioutil.ReadAll`. The backing slice for this buffer has a minimum size
of 512 bytes, even for small messages, because of the minimum read-size
in the stdandard library.

- A byte slice with approximately the same size as the generated
ProtoBuf response, allocated as part of `proto.Marshal`.

Both these memory allocations are used as scratch buffers, independently
of the allocation sizes for the generated ProtoBuf request and response
objects.

We have found that these memory allocations are significant enough to
show up when profiling Twirp services that deal with large or complex
ProtoBuf schemas, in either request or response form. Since scratch
buffers will always be required to parse and serialize ProtoBuf objects
in Go, this commit attempts to minimize the impact of the buffers by
using a memory pool to re-use them between requests.

These memory pools have been implemented in a backwards-compatible way
in the `serveXXXProtobuf` methods, and are disabled by default. An user
of the generated Twirp code may call `SetBufferPools` in the generated
server instance to set the memory pool objects that will be used to
serialize requests and responses; a different pool may be set for
request and for response buffers, or pooling may be disabled for either.

A new interface, `BufferPool` has been provided as part of the codegen
code. This is the interface that the memory pools must fulfill. The
standard library's `sync.Pool` fulfills this interface, so it's
straightforward to simply pass a `Pool` when configuring pooling, but
more advanced users may want to implement more complex logic for memory
pooling and/or eviction.

Because serialization and de-serialization in ProtoBuf use different
APIs, the actual implementations for each case have been slightly
modified so they can be consistently abstracted using the same memory
pool and so the functionality is equivalent to the existing code when
Pooling is disabled.

- For de-serialization, we need to read the whole contents of an
`io.Reader` into a byte slice. The old version of the code used
`ioutil.ReadAll`, which allocates a `bytes.Buffer` behind the scenes.
For this new version, we explicitly create an empty `bytes.Buffer` when
pooling is disabled and call `ReadFrom` on it, so this is functionally
equivalent to how de-serialization worked before the commit. When
pooling is enabled, we also create a `bytes.Buffer`, but we set the
backing slice to a pooled zero-size byte slice. This will prevent
`ReadFrom` from growing the backing slice as long as the pooled instance
has capacity. If the pooled instance does not have enough capacity, it
will be grown and cleaned up by the GC, and we will return the newly
allocated backing slice to the pool.

- For serialization, we do not call `proto.Marshal` directly in
any case, because this API will always allocate a new buffer. Instead,
we stack allocate a `proto.Buffer`. This struct is provided by the
stable proto API and allows us to use its `Marshal` method to serialize
an object directly into it. This is functionally equivalent to
`proto.Marshal` for the unpooled case, but when pooling is enabled, we
can replace the backing slice of the buffer with a byte slice from the
buffer pool. The same semantics as de-serialization apply here: as long
as there's capacity in the backing slice it won't be expanded, greatly
reducing the average allocation size.

It is worth noting that both (de)serialization cases have also been
carefully designed to trigger the most efficient code when your ProtoBuf
schemas have been generated using the alternative Gogo ProtoBuf
compiler.

To verify the impact of these optimizations, a set of benchmarks has
been designed where an Echo Twirp server processes complex ProtoBuf
messages. For this case, `Bench1` is a complex multi-field message from
Google's ProtoBuf benchmark suite, and `Bench2` is a large-size message
taken from one of GitHub's internal services.

Each message type is benchmarked using the default ProtoBuf
(de)serialization code and the Gogo ProtoBuf codegen. The results are as
follows (old is the old Twirp codgen; new is the codegen with pooling
enabled):

benchmark                  old ns/op     new ns/op     delta
BenchmarkBench1-16         10030         8126          -18.98%
BenchmarkBench1Gogo-16     7954          5324          -33.07%
BenchmarkBench2-16         5598061       4093236       -26.88%
BenchmarkBench2Gogo-16     2989519       1905072       -36.27%

benchmark                  old allocs     new allocs     delta
BenchmarkBench1-16         35             34             -2.86%
BenchmarkBench1Gogo-16     36             34             -5.56%
BenchmarkBench2-16         169            158            -6.51%
BenchmarkBench2Gogo-16     169            158            -6.51%

benchmark                  old bytes     new bytes     delta
BenchmarkBench1-16         5264          2386          -54.67%
BenchmarkBench1Gogo-16     9840          2610          -73.48%
BenchmarkBench2-16         8492147       2192365       -74.18%
BenchmarkBench2Gogo-16     8492152       2142307       -74.77%

To note: although the results are very significant for both message
types, using the Gogo ProtoBuf (de)serializer with memory pooling
provides the greatest gains. In any case, the benefit of enabling memory
pooling is always greater than the benefit of switching to Gogo
ProtoBuf, and using both always compounds.
```