@vmg 2019-03-25
@spenczar: cool. I'll follow up on the other issue with some notes about the different alternatives you're suggesting. It may take me a few days though. :bow: