@spenczar 2019-03-25
@vmg Sure - no need to do it all in one super-response. If you can provide details about what was slow specifically, that would help tremendously. You mentioned that deserialization showed up in profiles, but I'd like to know whether you mean CPU profiles, or memory profiles.